+++ src/main/java/com/atlauncher/constants/UIConstants.java
@@ -34,4 +34,8 @@ public class UIConstants {
     public static final Insets CHECKBOX_FIELD_INSETS = new Insets(SPACING_LARGE, -SPACING_SMALL, SPACING_LARGE, 0);
     public static final Insets CHECKBOX_FIELD_INSETS_SMALL = new Insets(SPACING_SMALL, -SPACING_SMALL, SPACING_SMALL,
             0);
+
+    // When using FlowLayout with a horizonal margin, we need to negate the first
+    // components margin added from the FlowLayout
+    public static final Insets FLOW_FIELD_INSETS = new Insets(SPACING_LARGE, (-SPACING_LARGE) - 3, SPACING_LARGE, 0);
 }